+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/cases/definition/repository/CaseDefinitionJpaRepositoryImpl.java
@@ -69,7 +69,6 @@ public class CaseDefinitionJpaRepositoryImpl implements CaseDefinitionRepository
         existingEntity.setDeployed(caseDefinition.getDeployed());
         existingEntity.setStages(caseDefinition.getStages());
         existingEntity.setCaseHooks(caseDefinition.getCaseHooks());
-        existingEntity.setKanbanConfig(caseDefinition.getKanbanConfig());
 
         entityManager.merge(existingEntity);
     }