+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/jpa/entity/CaseDefinitionEntity.java
@@ -3,14 +3,12 @@ package com.wks.caseengine.jpa.entity;
 import java.util.List;
 import java.util.UUID;
 
-import com.google.gson.JsonObject;
 import com.wks.caseengine.cases.definition.CaseStage;
 import com.wks.caseengine.cases.definition.DocumentRequirement;
 import com.wks.caseengine.event.ActionHook;
 import com.wks.caseengine.jpa.entity.converter.ActionHookListConverter;
 import com.wks.caseengine.jpa.entity.converter.CaseStageListConverter;
 import com.wks.caseengine.jpa.entity.converter.DocumentRequirementListConverter;
-import com.wks.caseengine.jpa.entity.converter.JsonConverter;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
@@ -58,9 +56,5 @@ public class CaseDefinitionEntity {
 	@Column(name="required_documents", columnDefinition = "TEXT")
 	@Convert(converter = DocumentRequirementListConverter.class)
 	private List<DocumentRequirement> requiredDocuments;
-
-	@Column(name="kanban_config", columnDefinition = "TEXT")
-	@Convert(converter = JsonConverter.class)
-	private JsonObject kanbanConfig;
 
 }