+++ apps/react/case-portal/src/views/management/caseDef/caseDefForm/caseDefForm.js
@@ -16,7 +16,6 @@ import { CaseDefFormEvents } from './listeners/caseDefFormListenerList'
 import { CaseDefFormStages } from './caseDefFormStages'
 import { CaseDefGeneralForm } from './caseDefGeneralForm'
 import { CaseDefFormForm } from './caseDefFormForm'
-import { CaseKanbanForm } from './caseDefKanban'
 import { CaseDefFormDocuments } from './caseDefFormDocuments'
 import { CaseDefSchemaView } from './caseDefSchemaView'
 import { CaseDefService, MenuEventService } from 'services'
@@ -151,13 +150,12 @@ export const CaseDefForm = ({ open, handleClose, caseDefParam }) => {
             <Tab label='General' {...a11yProps(0)} />
             <Tab label='Stages' {...a11yProps(1)} />
             <Tab label='Event Listeners' {...a11yProps(2)} />
-            <Tab label='Kanban' {...a11yProps(3)} />
-            <Tab label='Documents' {...a11yProps(4)} />
+            <Tab label='Documents' {...a11yProps(3)} />
             <Tab
               label={
                 schemaErrorCount > 0 ? `JSON (${schemaErrorCount})` : 'JSON'
               }
-              {...a11yProps(5)}
+              {...a11yProps(4)}
             />
           </Tabs>
         </Box>
@@ -179,14 +177,10 @@ export const CaseDefForm = ({ open, handleClose, caseDefParam }) => {
         </TabPanel>
 
         <TabPanel value={tabValue} index={3}>
-          <CaseKanbanForm caseDef={caseDef} setCaseDef={setCaseDef} />
-        </TabPanel>
-
-        <TabPanel value={tabValue} index={4}>
           <CaseDefFormDocuments caseDef={caseDef} setCaseDef={setCaseDef} />
         </TabPanel>
 
-        <TabPanel value={tabValue} index={5}>
+        <TabPanel value={tabValue} index={4}>
           <CaseDefSchemaView caseDef={caseDef} />
         </TabPanel>
       </Dialog>